index works, needed constructor for itemDAO and closing tag for img in view

+++ src/main/java/com/codeup/lostfound/repositories/ItemRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ItemRepository extends CrudRepository <Item, Long>  {
+public interface ItemRepository extends CrudRepository <Item, Integer>  {
 
     List<Item>findAll();
+
 
 
 //
